changed all the decimals to floats

+++ backend/db/migrations/20230709011522-create-event.js
@@ -50,7 +50,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.DECIMAL(6,2)
+        type: Sequelize.FLOAT
       },
       startDate: {
         type: Sequelize.DATE,
